AG-13489 Remove unused InteractionManager code
