+++ packages/ag-charts-community/src/chart/interaction/interactionManager.ts
@@ -1,195 +1,18 @@
-import type { DOMManager } from '../../dom/domManager';
-import { Debug } from '../../util/debug';
-import { getWindow } from '../../util/dom';
-import type { Listeners } from '../../util/listeners';
-import { partialAssign } from '../../util/object';
-import { isFiniteNumber } from '../../util/type-guards';
-import { InteractionState, InteractionStateListener } from './interactionStateListener';
-import { type PreventableEvent, type Unpreventable, buildPreventable, dispatchTypedEvent } from './preventableEvent';
-
-export { InteractionState };
-
-export const POINTER_INTERACTION_TYPES = [
-    'click',
-    'dblclick',
-    'contextmenu',
-    'hover',
-    'leave',
-    'enter',
-    'page-left',
-    'wheel',
-] as const;
-
-const KEY_INTERACTION_TYPES = ['keydown', 'keyup'] as const;
-
-export type PointerInteractionTypes = (typeof POINTER_INTERACTION_TYPES)[number];
-
-export type KeyInteractionTypes = (typeof KEY_INTERACTION_TYPES)[number];
-
-export type InteractionTypes = PointerInteractionTypes | KeyInteractionTypes;
-
-type SUPPORTED_EVENTS =
-    | 'dblclick'
-    | 'contextmenu'
-    | 'keydown'
-    | 'keyup'
-    | 'mousedown'
-    | 'mousemove'
-    | 'mouseup'
-    | 'mouseleave'
-    | 'mouseenter'
-    | 'touchstart'
-    | 'touchmove'
-    | 'touchend'
-    | 'touchcancel'
-    | 'pagehide'
-    | 'wheel';
-const SHADOW_DOM_HANDLERS: SUPPORTED_EVENTS[] = ['mousemove', 'mouseup'];
-const WINDOW_EVENT_HANDLERS: SUPPORTED_EVENTS[] = ['pagehide', 'mousemove', 'mouseup'];
-const EVENT_HANDLERS = [
-    'click',
-    'dblclick',
-    'contextmenu',
-    'mousedown',
-    'mouseleave',
-    'mouseenter',
-    'touchstart',
-    'touchmove',
-    'touchend',
-    'touchcancel',
-    'wheel',
-    'keydown',
-    'keyup',
-] as const;
-
-type BaseInteractionEvent<T extends InteractionTypes, TEvent extends Event> = PreventableEvent & {
-    type: T;
-    sourceEvent: TEvent;
-    relatedElement?: HTMLElement;
-    targetElement?: HTMLElement;
-};
-
-export type PointerOffsets = {
-    offsetX: number;
-    offsetY: number;
-    pageX?: number;
-    pageY?: number;
-};
-
-export type PointerHistoryEvent = PointerOffsets & { type: string };
-
-export type PointerInteractionEvent<T extends PointerInteractionTypes = PointerInteractionTypes> = PointerOffsets &
-    BaseInteractionEvent<T, Event> & {
-        pageX: number;
-        pageY: number;
-        deltaX: number;
-        deltaY: number;
-        button: number;
-        pointerHistory: PointerHistoryEvent[];
-    };
-
-export type KeyInteractionEvent<T extends KeyInteractionTypes = KeyInteractionTypes> = BaseInteractionEvent<
-    T,
-    KeyboardEvent
->;
-
-export type InteractionEvent =
-    | PointerInteractionEvent<PointerInteractionTypes>
-    | KeyInteractionEvent<KeyInteractionTypes>;
-
-interface Coords {
-    clientX: number;
-    clientY: number;
-    pageX: number;
-    pageY: number;
-    offsetX: number;
-    offsetY: number;
+// These interaction state are both bitflags and priorities.
+// Smaller numbers have higher priority, because it is possible to find the least
+// significant bit in O(1) complexity using a bitwise operation.
+export enum InteractionState {
+    Default = 32,
+    ZoomDrag = 16,
+    Annotations = 8,
+    ContextMenu = 4,
+    Animation = 2,
+    AnnotationsSelected = 1,
+    All = Default | ZoomDrag | Annotations | ContextMenu | Animation | AnnotationsSelected,
 }
 
-type SupportedEvent = MouseEvent | TouchEvent | Event;
-
-function isPointerEvent(type: InteractionTypes): type is PointerInteractionTypes {
-    return POINTER_INTERACTION_TYPES.includes(type as any);
-}
-
-function isKeyEvent(type: InteractionTypes): type is KeyInteractionTypes {
-    return KEY_INTERACTION_TYPES.includes(type as any);
-}
-
-/**
- * Manages user interactions with a specific HTMLElement (or interactions that bubble from it's
- * children)
- */
-export class InteractionManager extends InteractionStateListener<InteractionTypes, InteractionEvent> {
-    private readonly debug = Debug.create(true, 'interaction');
-
-    private rootElement: HTMLElement | undefined;
-
-    private readonly eventHandler = (event: SupportedEvent) => this.processEvent(event);
-    private readonly clickHistory: [PointerHistoryEvent] = [{ offsetX: NaN, offsetY: NaN, type: 'mousedown' }];
-    private readonly dblclickHistory: [PointerHistoryEvent, PointerHistoryEvent, PointerHistoryEvent] = [
-        { offsetX: NaN, offsetY: NaN, type: 'mousedown' },
-        { offsetX: NaN, offsetY: NaN, type: 'mouseup' },
-        { offsetX: NaN, offsetY: NaN, type: 'mousedown' },
-    ];
-
+export class InteractionManager {
     private stateQueue: InteractionState = InteractionState.Default | InteractionState.Animation;
-
-    public constructor(
-        private readonly keyboardOptions: { readonly enabled: boolean },
-        private readonly domManager: DOMManager
-    ) {
-        super();
-
-        this.rootElement = this.domManager.getShadowDocumentRoot();
-
-        for (const type of EVENT_HANDLERS) {
-            if (type.startsWith('touch') || type === 'wheel') {
-                this.domManager.addEventListener(type, this.eventHandler, { passive: false });
-            } else {
-                this.domManager.addEventListener(type, this.eventHandler);
-            }
-        }
-
-        for (const type of WINDOW_EVENT_HANDLERS) {
-            getWindow().addEventListener(type, this.eventHandler);
-        }
-
-        this.containerChanged(true);
-        this.domManager.addListener('container-changed', () => this.containerChanged());
-    }
-
-    private containerChanged(force = false) {
-        const newRoot = this.domManager.getShadowDocumentRoot();
-        if (!force && newRoot === this.rootElement) return;
-
-        for (const type of SHADOW_DOM_HANDLERS) {
-            this.rootElement?.removeEventListener(type, this.eventHandler);
-        }
-
-        this.rootElement = newRoot;
-        this.debug('[InteractionManager] Switching rootElement to:', this.rootElement);
-
-        for (const type of SHADOW_DOM_HANDLERS) {
-            this.rootElement?.addEventListener(type, this.eventHandler);
-        }
-    }
-
-    override destroy() {
-        super.destroy();
-
-        for (const type of WINDOW_EVENT_HANDLERS) {
-            getWindow().removeEventListener(type, this.eventHandler);
-        }
-        for (const type of SHADOW_DOM_HANDLERS) {
-            this.rootElement?.removeEventListener(type, this.eventHandler);
-        }
-        for (const type of EVENT_HANDLERS) {
-            this.domManager.removeEventListener(type, this.eventHandler);
-        }
-
-        this.domManager.removeStyles('interactionManager');
-    }
 
     public pushState(state: InteractionState) {
         this.stateQueue |= state;
@@ -199,183 +22,9 @@ export class InteractionManager extends InteractionStateListener<InteractionType
         this.stateQueue &= ~state;
     }
 
-    public override getState(): InteractionState {
+    public getState(): InteractionState {
         // Bitwise operation to get the least significant bit:
         return this.stateQueue & -this.stateQueue;
-    }
-
-    private processEvent(event: SupportedEvent) {
-        this.debug('Received raw event', event);
-
-        let types = this.decideInteractionEventTypes(event);
-        if (types != null && !Array.isArray(types)) {
-            types = [types];
-        }
-
-        for (const type of types ?? []) {
-            // Async dispatch to avoid blocking the event-processing thread.
-            this.dispatchEvent(event, type);
-        }
-    }
-
-    private dispatchEvent(event: SupportedEvent, type: InteractionTypes) {
-        if (isPointerEvent(type)) {
-            this.dispatchPointerEvent(event, type);
-            return;
-        }
-
-        const { relatedElement, targetElement } = this.extractElements(event);
-        if (isKeyEvent(type)) {
-            const sourceEvent = event as KeyboardEvent;
-            this.dispatchTypedEvent(this.listeners, { type, sourceEvent, relatedElement, targetElement });
-        }
-    }
-
-    private dispatchTypedEvent<
-        T extends string,
-        E extends { type: T },
-        L extends Listeners<T, (event: PreventableEvent & E) => void>,
-    >(listeners: L, event: E) {
-        const preventableEvent = buildPreventable(event);
-        this.debug('Dispatching typed event', preventableEvent, this.getState());
-        listeners.dispatchWrapHandlers(event.type, (handler, e) => handler(e), preventableEvent);
-    }
-
-    extractElements(event: SupportedEvent): { relatedElement?: HTMLElement; targetElement?: HTMLElement } {
-        let relatedElement;
-        let targetElement;
-
-        if ('relatedTarget' in event && event['relatedTarget'] instanceof HTMLElement) {
-            relatedElement = event['relatedTarget'];
-        }
-        if ('target' in event && event['target'] instanceof HTMLElement) {
-            targetElement = event['target'];
-        }
-
-        return { relatedElement, targetElement };
-    }
-
-    private dispatchPointerEvent(event: SupportedEvent, type: PointerInteractionTypes) {
-        const coords = this.calculateCoordinates(event);
-        if (coords == null) return;
-
-        const pointerEvent = this.buildPointerEvent({ type, event, ...coords });
-        this.debug('Dispatching pointer event', pointerEvent, this.getState());
-        dispatchTypedEvent(this.listeners, pointerEvent);
-    }
-
-    private getEventHTMLTarget(event: SupportedEvent): HTMLElement | undefined {
-        if (event.target instanceof HTMLElement) {
-            return event.target;
-        } else if (event.currentTarget instanceof HTMLElement) {
-            return event.currentTarget;
-        }
-        return undefined;
-    }
-
-    private recordDown(event: SupportedEvent) {
-        if (event instanceof MouseEvent) {
-            partialAssign(['offsetX', 'offsetY'], this.clickHistory[0], event);
-            partialAssign(['offsetX', 'offsetY'], this.dblclickHistory[2], this.dblclickHistory[0]);
-            partialAssign(['offsetX', 'offsetY'], this.dblclickHistory[0], event);
-        }
-    }
-
-    private recordUp(event: SupportedEvent) {
-        if (event instanceof MouseEvent) {
-            partialAssign(['offsetX', 'offsetY'], this.dblclickHistory[1], event);
-        }
-    }
-
-    private decideInteractionEventTypes(event: SupportedEvent): InteractionTypes | InteractionTypes[] | undefined {
-        switch (event.type) {
-            case 'keydown':
-            case 'keyup':
-                return this.keyboardOptions.enabled ? event.type : undefined;
-
-            case 'click':
-            case 'dblclick':
-                return event.type;
-
-            case 'contextmenu':
-            case 'wheel':
-                return event.type;
-
-            case 'touchstart':
-            case 'mousedown':
-                if (!this.isEventOverElement(event)) {
-                    return;
-                }
-                this.recordDown(event);
-                return;
-
-            case 'touchmove':
-            case 'mousemove':
-                if (!this.isEventOverElement(event)) {
-                    return;
-                }
-                return 'hover';
-
-            case 'mouseup':
-            case 'touchend':
-                if (!this.isEventOverElement(event)) {
-                    return;
-                }
-                this.recordUp(event);
-                return;
-
-            case 'mouseleave':
-            case 'touchcancel':
-                return 'leave';
-
-            case 'mouseenter':
-                return 'enter';
-
-            case 'pagehide':
-                return 'page-left';
-        }
-    }
-
-    private isEventOverElement(event: SupportedEvent) {
-        return event.target instanceof HTMLElement && this.domManager.contains(event.target);
-    }
-
-    private static readonly NULL_COORDS: Coords = {
-        clientX: -Infinity,
-        clientY: -Infinity,
-        pageX: -Infinity,
-        pageY: -Infinity,
-        offsetX: -Infinity,
-        offsetY: -Infinity,
-    };
-
-    private calculateCoordinates(event: SupportedEvent): Coords | undefined {
-        if (event instanceof MouseEvent) {
-            return this.getMouseEventCoords(event);
-        } else if (typeof TouchEvent !== 'undefined' && event instanceof TouchEvent) {
-            const lastTouch = event.touches[0] ?? event.changedTouches[0];
-            const { clientX, clientY, pageX, pageY } = lastTouch;
-            return { ...InteractionManager.NULL_COORDS, clientX, clientY, pageX, pageY };
-        } else if (event instanceof PageTransitionEvent) {
-            if (event.persisted) {
-                // Don't fire the page-left event since the page maybe revisited.
-                return;
-            }
-            return InteractionManager.NULL_COORDS;
-        }
-
-        // Unsupported event - abort.
-    }
-
-    private getMouseEventCoords(event: MouseEvent): Coords {
-        const { clientX, clientY, pageX, pageY } = event;
-        let { offsetX, offsetY } = event;
-
-        const target = this.getEventHTMLTarget(event);
-        const { x = 0, y = 0 } = target ? this.domManager.calculateCanvasPosition(target) : {};
-        offsetX += x;
-        offsetY += y;
-        return { clientX, clientY, pageX, pageY, offsetX, offsetY };
     }
 
     private static isWheelEvent(event: Event): event is WheelEvent {
@@ -392,63 +41,5 @@ export class InteractionManager extends InteractionStateListener<InteractionType
             deltaY = event.deltaY * factor;
         }
         return { deltaX, deltaY };
-    }
-
-    private buildPointerEvent(opts: {
-        type: PointerInteractionTypes;
-        event: Event;
-        clientX: number;
-        clientY: number;
-        offsetX?: number;
-        offsetY?: number;
-        pageX?: number;
-        pageY?: number;
-    }): Unpreventable<PointerInteractionEvent<PointerInteractionTypes>> {
-        const { type, event, clientX, clientY } = opts;
-        let { offsetX, offsetY, pageX, pageY } = opts;
-
-        if (!isFiniteNumber(offsetX) || !isFiniteNumber(offsetY)) {
-            const rect = this.domManager.getBoundingClientRect();
-            offsetX = clientX - rect.left;
-            offsetY = clientY - rect.top;
-        }
-        if (!isFiniteNumber(pageX) || !isFiniteNumber(pageY)) {
-            const pageRect = this.rootElement?.getBoundingClientRect();
-            pageX = clientX - (pageRect?.left ?? 0);
-            pageY = clientY - (pageRect?.top ?? 0);
-        }
-
-        const { deltaX, deltaY } = InteractionManager.getWheelDeltas(event);
-
-        // AG-8880 Because we are using listeners globally on the canvases, click events are always fired
-        // whenever the mouse button is lifted. The pointerHistory allows listeners to check that click events
-        // are only fired when both the mousedown & mouseup events are in the revelant bounding area.
-        let pointerHistory: PointerHistoryEvent[] = [];
-        if (event.type === 'click') {
-            pointerHistory = this.clickHistory;
-        } else if (event.type === 'dblclick') {
-            pointerHistory = this.dblclickHistory;
-        }
-
-        const { relatedElement, targetElement } = this.extractElements(event);
-
-        const button = 'button' in event ? Number(event.button) : 0;
-        const builtEvent = {
-            type,
-            offsetX,
-            offsetY,
-            pageX,
-            pageY,
-            deltaX,
-            deltaY,
-            button,
-            pointerHistory,
-            sourceEvent: event,
-            relatedElement,
-            targetElement,
-        };
-
-        this.debug('InteractionManager - builtEvent: ', builtEvent, this.getState());
-        return builtEvent;
     }
 }
